Phani | Move notes save & delete components into components directory
