+++ micro-frontends/src/next-ui/Components/otNotes/OtNotes.scss
