+++ micro-frontends/src/next-ui/index.js
@@ -1,8 +1,7 @@
 import { React2AngularBridgeBuilder } from "../utils/bridge-builder";
 import { PatientAlergiesControl } from "./Containers/patientAlergies/PatientAlergiesControl";
 import { FormDisplayControl } from "./Containers/formDisplayControl/FormDisplayControl";
-import { OtNotesSavePopup } from "./Containers/otNotes/OtNotesSavePopup";
-import { OtNotesDeletePopup } from "./Containers/otNotes/OtNotesDeletePopup";
+import { OtNotesSavePopup, OtNotesDeletePopup } from "./Containers/otNotes/OtNotes";
 
 const MODULE_NAME = "bahmni.mfe.nextUi";
